New loading ui working
Now i have to test it on prod and if everything works fine i will polish
the ui
