+++ package.json
@@ -18,28 +18,28 @@
     "@eslint/js": "^9.18.0",
     "@internationalized/date": "^3.8.2",
     "@lucide/svelte": "^0.515.0",
-    "@sveltejs/adapter-auto": "^4.0.0",
-    "@sveltejs/kit": "^2.16.0",
-    "@sveltejs/vite-plugin-svelte": "^5.0.0",
+    "@sveltejs/adapter-auto": "^6.1.0",
+    "@sveltejs/kit": "^2.37.0",
+    "@sveltejs/vite-plugin-svelte": "^6.1.4",
     "@tailwindcss/vite": "^4.1.7",
     "@types/node": "^22.15.21",
     "bits-ui": "^2.9.0",
     "clsx": "^2.1.1",
     "eslint": "^9.18.0",
     "eslint-config-prettier": "^10.0.1",
-    "eslint-plugin-svelte": "^3.0.0",
+    "eslint-plugin-svelte": "^3.12.0",
     "globals": "^16.0.0",
     "prettier": "^3.5.3",
     "prettier-plugin-svelte": "^3.3.3",
     "prettier-plugin-tailwindcss": "^0.6.11",
-    "svelte": "^5.0.0",
-    "svelte-check": "^4.0.0",
+    "svelte": "^5.38.6",
+    "svelte-check": "^4.3.1",
     "tailwind-merge": "^3.3.1",
     "tailwind-variants": "^1.0.0",
     "tw-animate-css": "^1.3.5",
     "typescript": "^5.0.0",
     "typescript-eslint": "^8.20.0",
-    "vite": "^6.2.5"
+    "vite": "^7.1.4"
   },
   "dependencies": {
     "@tanstack/table-core": "^8.21.3",
