+++ svelte.config.js
@@ -7,10 +7,19 @@ const config = {
   // for more information about preprocessors
   preprocess: vitePreprocess(),
 
+  compilerOptions: {
+    experimental: {
+      async: true
+    }
+  },
+
   kit: {
     // adapter-auto only supports some environments, see https://svelte.dev/docs/kit/adapter-auto for a list.
     // If your environment is not supported, or you settled on a specific environment, switch out the adapter.
     // See https://svelte.dev/docs/kit/adapters for more information about adapters.
+    experimental: {
+      remoteFunctions: true
+    },
     adapter: adapter()
   }
 };
